ai_travel_agent/requirements.txt
openai
python-dotenv
plotly